Removed all references to old runtime builder
